refactor: 자동차 이름을 별도의 클래스로 분리
자동차 이름 그리고 그에 대한 검증이 이루어지는 부분을 별도의 클래스로 분리하였습니다.
