+++ src/main/java/racingcar/domain/Car.java
@@ -4,19 +4,16 @@ public class Car {
 
     private static final int POSITION_INITIAL_VALUE = 0;
     private static final int PROGRESS_CONDITION_VALUE = 4;
-    private static final int CAR_NAME_MIN_LENGTH = 1;
-    private static final int CAR_NAME_MAX_LENGTH = 5;
 
-    private final String name;
+    private final CarName carName;
     private int position = POSITION_INITIAL_VALUE;
 
     public Car(String name) {
-        validateCarName(name);
-        this.name = name;
+        this.carName = new CarName(name);
     }
 
     public String getName() {
-        return name;
+        return carName.getName();
     }
 
     public int getPosition() {
@@ -34,31 +31,6 @@ public class Car {
             return true;
         }
         return false;
-    }
-
-    public void validateCarName(String carName) {
-        if (carName == null) {
-            throw new IllegalArgumentException("[ERROR] 자동차 이름은 null 일 수 없습니다.");
-        }
-
-        validateCarNameLength(carName);
-        validateCarNameSpace(carName);
-    }
-
-    private void validateCarNameLength(String carName) {
-        if (carName.length() < CAR_NAME_MIN_LENGTH) {
-            throw new IllegalArgumentException("[ERROR] 자동차 이름은 1글자 이상이어야 합니다.");
-        }
-
-        if (carName.length() > CAR_NAME_MAX_LENGTH) {
-            throw new IllegalArgumentException("[ERROR] 자동차 이름은 5글자 이하이어야 합니다.");
-        }
-    }
-
-    private void validateCarNameSpace(String carName) {
-        if (carName.contains(" ")){
-            throw new IllegalArgumentException("[ERROR] 공백 없이 입력해주세요.");
-        }
     }
 }
